try to simplify packaging, enforce tagged releases
package releasenum is supposed to be for different builds of the package
because of packaging changes, not for number of commits since the last
tag on the source, so generate the releasenum from the latest in the repo
for the software version being built. make the software version explicit
in the spec file, use that to build the source archive, rather than the
state of the current checkout.  generate source archive from a clean
checkout of the specified version, never from a dirty checkout.  the
packaged version and release can increment differently than the software
version, and a package cannot be released except from tagged source.