refactor installs
try to consolidate installation in scons not in spec, with sysconfig
installs under install.root alias.  generate nc_server.conf in the
source rather than editing the installed location.  replace
sharedlibrary tool with scons built-in InstallVersionedLib().  use
scons --install-sandbox to install into RPM_BUILD_ROOT without
redefining PREFIX.  BUILDS is now forced to be a single value and
not a list.  this commit depends on eol_scons changes which have
not been committed yet.  spec version can contain a ~ suffix for
pre-releases which will be converted to a dash in the version tag.