connect to NC_SERVER_PORT if set in environment
this uses a single implementation in an object file linkd into the clients.
it currently uses nidas::util to resolve the port address, same as the
NetcdfRPCChannel class in nidas, but it could be re-implemented without
nidas::util using just calls to getaddrinfo() and such.  the new client
functions could be included in the nc_server_rpc library and used by
NetcdfRPCChannel also at some point, so far they are not in that library, so
as not to introduce a new dependency on nidas_util.  lastly, it might also
make sense to allow the port to be specified in the hostname, like
<server>:<port>, rather just in the environment.

the ParseConfig had been adding all of the nidas libraries as a dependency, so
now it has been fixed to only link against nidas_util, for both the server and
the client programs.