use get_att_unique<> in global scope
NS_NcVar does not inherit from NcVar, but NS_NcFile does inherit from NcFile,
and nc_server needs to call get_att on pointers to all of them, so create a
template function which calls get_att() on any of those types and returns a
unique_ptr<NcAtt>.  Code which was calling get_att() on NcVar* was resolving
to the non-unique_ptr version of get_att and still leaking, so now calling
get_att_unique() makes it explicit that a unique_ptr should be returned.

also remove an OutVariable constructor declaration that was neither used nor
defined.