setup parallel 8 and 9 builds, skip local update
local package updates on 8 and 9 servers are not working at the moment, so
skip for nc-server same as for nidas

trigger rebuilds of nc-server packages when NIDAS is rebuilt, except this is
not quite correct.  since the local package updates are not working, nc-server
will be rebuilt against the nidas package currently installed on the server,
rather than the latest nidas package.  a container could fix this i suppose...
omit vortex32 build since nc-server is not needed on that target.