store attributes from the processed variables
if a variable name can be matched up with a processed variable from the
Project (ie, a Variable attached to a DSMSensor and processed by it), then add
that Variable's attributes to the netcdf variable.  this depends on the nidas
changes on the m2hats-allow-heights branch which are not yet merged to master.

even though attributes are copied to variables in derived sample tags, the
derived variables lose the reference to their input variables, so derived
variables would not see any changes to the attributes in the source variables.
it might make sense someday to track those references and allow the changes to
propagate automatically.

also create a CStringCache class to cache char arrays and delete them when the
cache instance goes out of scope.  this seems like a more robust way to ensure
that the memory referenced in the xdr structures remains valid, rather than
depending on other code to keep std::string memory valid.